% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{adj_lst}
\alias{adj_lst}
\alias{adj_lst.gengraph}
\title{Adjacency List}
\usage{
adj_lst(x)

\method{adj_lst}{gengraph}(x)
}
\arguments{
\item{x}{\code{gengraph}}
}
\value{
An adjacency list
}
\description{
Extracts the adjacency list of a \code{gengraph}
}
