% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{adj_mat}
\alias{adj_mat}
\alias{adj_mat.gengraph}
\title{Adjacency Matrix}
\usage{
adj_mat(x)

\method{adj_mat}{gengraph}(x)
}
\arguments{
\item{x}{\code{gengraph} object}
}
\value{
An adjacency matrix
}
\description{
Extracts the adjacency matrix of a \code{gengraph} object
}
