% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{as_adj_lst}
\alias{as_adj_lst}
\title{Converts an adjacency matrix to an adjacency list}
\usage{
as_adj_lst(A)
}
\arguments{
\item{A}{Adjacency matrix}
}
\description{
Converts an adjacency matrix to an adjacency list
}
