% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{as_adj_mat}
\alias{as_adj_mat}
\title{Converts an adjacency list to an adjacency matrix}
\usage{
as_adj_mat(adj)
}
\arguments{
\item{adj}{Adjacency list}
}
\value{
An adjacency matrix
}
\description{
Converts an adjacency list to an adjacency matrix
}
\examples{
adj <- list(a = c("b", "d"), b = c("a", "c", "d"), c = c("b", "d"), d = c("a", "c", "b"))
as_adj_mat(adj)
}
