% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.gengraph}
\title{Gengraph as igraph}
\usage{
as_igraph(x)

\method{as_igraph}{gengraph}(x)
}
\arguments{
\item{x}{\code{gengraph} object}
}
\value{
An \code{igraph} object
}
\description{
Convert a \code{gengraph} object to an \code{igraph} object
}
