% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_decomp_graphs.R
\name{components}
\alias{components}
\title{Finds the components of a graph}
\usage{
components(adj)
}
\arguments{
\item{adj}{Adjacency list or \code{gengraph} object}
}
\value{
A list where the elements are the components of the graph
}
\description{
Finds the components of a graph
}
