% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dfs}
\alias{dfs}
\title{Depth First Search}
\usage{
dfs(adj, root)
}
\arguments{
\item{adj}{A named adjacency list of a decomposable grah}

\item{root}{The node from which the component should be found}
}
\value{
All nodes connected to \code{root}
}
\description{
Finds the elements in the component of \code{root}
}
\examples{
x <- list(a = c("b", "d"), b = c("a", "d"), c = c("b", "a"),
          d = c("e", "f"), e = c("d", "f"), f = c("d", "e"))
dfs(x, "a")
}
