\name{ESS-package}
\alias{ESS-package}
\alias{ESS}
\docType{package}
\title{
\packageTitle{ESS}
}
\description{
\packageDescription{ESS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ESS}
\packageIndices{ESS}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ESS}

Maintainer: \packageMaintainer{ESS}
}
\references{
Wiesenfarth, M., Calderazzo, S. (2019). Quantification of Prior Impact in Terms of Effective Current Sample Size. Submitted.

}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{

%}
