% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_decomp_graphs.R
\name{make_null_graph}
\alias{make_null_graph}
\title{Make a null graph}
\usage{
make_null_graph(nodes)
}
\arguments{
\item{nodes}{A character vector containing the nodes to be used in the graph}
}
\value{
An adjacency list the null graph with no edges
}
\description{
A helper function to make an adjacency list corresponding to a null graph (no edges)
}
\examples{
d  <- derma[, 5:8]
ng <- make_null_graph(colnames(d))
}
