% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcs}
\alias{mcs}
\title{Maximum Cardinality Search}
\usage{
mcs(adj, check = TRUE)
}
\arguments{
\item{adj}{A named adjacency list of a decomposable graph}

\item{check}{Boolean: check if adj is decomposable}
}
\value{
A list with a perfect numbering of the nodes and a perfect sequence of sets
}
\description{
Maximum Cardinality Search
}
\details{
If adj is not the adjacency list of a decomposable graph an error is raised
}
\examples{
x <- list(a = c("b", "d"), b = c("a", "c", "d"), c = c("b", "d"), d = c("a", "c", "b"))
mcs(x)
}
