% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{plot.gengraph}
\alias{plot.gengraph}
\title{Plot}
\usage{
\method{plot}{gengraph}(x, vc = NULL, ...)
}
\arguments{
\item{x}{A \code{gengraph} object}

\item{vc}{Named character vector; the names are the vertices and
the elements are the colors of the nodes}

\item{...}{Extra arguments. See the igraph package}
}
\value{
No return value, called for side effects
}
\description{
A wrapper around igraphs plot method for \code{gengraph} objects
}
\examples{

d <- derma[, 10:25]
g <- fit_graph(d)
vs <- colnames(d)
vcol <- structure(vector("character", length(vs)), names = vs)
vcol[1:4]  <- "lightsteelblue2"
vcol[5:7]  <- "orange"
vcol[8:16] <- "pink"
plot(g, vcol)
}
