% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gengraph.R
\name{print.gengraph}
\alias{print.gengraph}
\title{Print}
\usage{
\method{print}{gengraph}(x, ...)
}
\arguments{
\item{x}{A \code{gengraph} object}

\item{...}{Not used (for S3 compatability)}
}
\description{
A print method for \code{gengraph} objects
}
