% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_entropy.R
\name{entropy}
\alias{entropy}
\title{Joint Entropy}
\usage{
entropy(df, thres = 5)
}
\arguments{
\item{df}{data.frame}

\item{thres}{A threshold mechanism for choosing between two different ways of calculating the entropy. Can Speed up the procedure with the "correct" value.}
}
\value{
A number representing the entropy of the variables in \code{df}.
}
\description{
Calculates the joint entropy over discrete variables in \code{df}
}
\examples{
entropy(derma[1:100, 1:3])

}
