% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\docType{package}
\name{ess-package}
\alias{ess}
\alias{ess-package}
\title{ess: Eficient Stepwise Selection in Decomposable Models}
\description{
The class of graphical models is a family of probability
distributions for which conditional dependencies can be read off
from a graph. If the graph is decomposable, the maximum
likelihood estimates of the parameters in the model can be shown
to be on exact form. This is what enables ESS to be fast and
efficient for model selection in decomposable graphical models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlindsk/ess}
  \item Report bugs at \url{https://github.com/mlindsk/ess/issues}
}

}
\author{
\strong{Maintainer}: Mads Lindskou \email{mads@math.aau.dk}

}
