\name{ESS-package}
\alias{ESS-package}
\alias{ESS}
\docType{package}
\title{
\packageTitle{ESS}
}
\description{
\packageDescription{ESS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ESS}
\packageIndices{ESS}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ESS}

Maintainer: \packageMaintainer{ESS}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%~~ directory ~~
\note{This is a preliminary version and changes in syntax are not unlikely.
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{

%}
