% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gengraph.R
\name{gengraph}
\alias{gengraph}
\title{A generic and extendable structure for decomposable graphical models}
\usage{
gengraph(df, type = "gen", adj = NULL, q = 0.5, ...)
}
\arguments{
\item{df}{data.frame}

\item{type}{character ("fwd", "bwd", "tree", "tfwd", "gen")}

\item{adj}{A user-specified adjacency list}

\item{q}{Penalty term in the stopping criterion (\code{0} = AIC and \code{1} = BIC)}

\item{...}{Not used (for extendibility)}
}
\value{
A \code{gengraph} object with child class \code{type} used for model selection.
}
\description{
A generic structure for decomposable graphical models
}
\examples{

gengraph(derma, type = "fwd")
gengraph(derma, type = "bwd")

}
\seealso{
\code{\link{adj_lst.gengraph}}, \code{\link{adj_mat.gengraph}}, \code{\link{fit_graph}}, \code{\link{walk.fwd}}, \code{\link{walk.bwd}}
}
